Journey Planner for Paris Metro
The name of the project is Paris Metro System.The codes of the project are written in Java.The project was made from November 13, 2023 to January 4, 2024.  We are expected to develop an algorithm to find the best journey between the given origin and destination stops. We should consider two optimization criteria: fewer stops and minimum time. In the first case, we should use equal edge weights in the graph. In the second case, we should use the given time intervals in the consecutive stops. We should represent each metro station as a node and represent each line connecting two consecutive stops in a certain direction as a directed edge to form a transportation graph. To make point to point queries in a transportation network, some sort of walk(transfer) edges are required, so any stage of the journey can be covered by walk or passengers may walk between the stops while transferring between two different lines. Walk-distance edges are also providing to link each of the transportation networks (bus, train, metro, ferry etc.). Two stops u and v are labeled as neighbor stops by adding walk-distance edges between them, if a road segment is available to pedestrians and dist(u,v) is less than the maximum allowed walking distance.